First version
Today the work with REST api and respond in JSON, XML has been in
ancient times for web services.

This PHP class helps create error type answers or responses to various
types of data and returning a JSON string.

In the folder there are files such example of how to implement an API
and a client.